Completed the Design
The tool now works as intended and the code is quite clean.

I'm using a library called WindowsInput to type the text into the active windows and I added a GlobalKeyboardHook event to listen to whenever a hotkey is pressed.

On top of that, I've also added FodyWeavers to compile the app into a single executable (merges the WindowsInput.dll into the .exe file).

I also want to make it so it saves the text that was written in the Text Boxes instead of always being empty whenever you launch the app again.