Added an auto save
Text within the text boxes will now be saved. Meaning that you can close the app and reopen it and the text will still be there. Also added an hyperlink on the About panel to link my website.

Next thing I plan on adding is something to prevent 2 instances of the app to be running.